印刷処理の改善と設定の型変更
RePrintWindow.cs では、印刷メソッドを `StartPrint` から `Print` に変更し、印刷処理のロジックを改善しました。印刷時の単位をインチからミリメートルに変更し、ローディング画面の表示を統合しました。また、ラベルとバーコードの印刷設定の計算精度を向上させました。

ProductBarcodeSettingsWindow.Designer.cs では、バーコード設定の UI コンポーネント名を変更しました。

ProductPrintLabelSetting.cs では、ラベルページ設定のプロパティを `double` 型から `float` 型に変更し、精度を向上させました。

ProductRegistration2Window.cs でも印刷メソッドを変更し、印刷設定の計算をインチからミリメートルに変更しました。

SubstratePrintSettingsWindow.cs では、サブストレートの印刷設定のプロパティを `double` 型から `float` 型に変更しました。

SubstrateRegistrationWindow.cs では、ラベル画像生成のサイズ計算を `decimal` 型から `int` 型に変更し、メモリ管理を効率化しました。テキスト描画のロジックも改善され、より正確な位置に描画されるようになりました。

+++ ProductDataBase/Substrate/SubstratePrintSetting.cs
@@ -16,15 +16,15 @@ namespace ProductDatabase.Substrate {
     public class LabelPageSettings {
         public int LabelsPerRow { get; set; }
         public int LabelsPerColumn { get; set; }
-        public double LabelWidth { get; set; }
-        public double LabelHeight { get; set; }
-        public double MarginX { get; set; }
-        public double MarginY { get; set; }
-        public double IntervalX { get; set; }
-        public double IntervalY { get; set; }
+        public float LabelWidth { get; set; }
+        public float LabelHeight { get; set; }
+        public float MarginX { get; set; }
+        public float MarginY { get; set; }
+        public float IntervalX { get; set; }
+        public float IntervalY { get; set; }
         public string HeaderTextFormat { get; set; } = string.Empty;
-        public double HeaderPositionX { get; set; }
-        public double HeaderPositionY { get; set; }
+        public float HeaderPositionX { get; set; }
+        public float HeaderPositionY { get; set; }
 
         [JsonConverter(typeof(FontJsonConverter))]
         public Font HeaderFont { get; set; } = SystemFonts.DefaultFont;
@@ -32,8 +32,8 @@ namespace ProductDatabase.Substrate {
 
     public class LabelLayoutSettings {
         public string TextFormat { get; set; } = string.Empty;
-        public double TextPositionX { get; set; }
-        public double TextPositionY { get; set; }
+        public float TextPositionX { get; set; }
+        public float TextPositionY { get; set; }
         public bool AlignTextXCenter { get; set; }
         public bool AlignTextYCenter { get; set; }
         public int CopiesPerLabel { get; set; }
